+++ importers/src/bible/osis/index.ts
@@ -13,17 +13,18 @@ import {
     IBibleContentGroup,
     IBibleContentPhrase,
     IBibleCrossReference,
-    IBibleNote,
+    IBibleNote
 } from '@bible-engine/core';
 
 import {
     endsWithNoSpaceAfterChar,
     getBibleReferenceFromParsedReference,
+    getBibleReferenceParserCustomBooks,
     getPhrasesFromParsedReferences,
     getReferencesFromText,
     isOnlyCrossReferenceWordOrPunctuation,
     startsWithNoSpaceBeforeChar,
-    streamToString,
+    streamToString
 } from '../../shared/helpers.functions';
 import { BibleEngineImporter, IImporterOptions } from '../../shared/Importer.interface';
 import Logger from '../../shared/Logger';
@@ -244,23 +245,36 @@ export class OsisImporter extends BibleEngineImporter {
                 )
                     context.version.crossRefBeforePhrase = true;
 
+                let bcv_parser;
                 try {
-                    const bcv_parser = require(`bible-passage-reference-parser/js/${context.version.language}_bcv_parser`)
+                    bcv_parser = require(`bible-passage-reference-parser/cjs/${context.version.language}_bcv_parser`)
                         .bcv_parser;
-                    const bcv: BibleReferenceParser = new bcv_parser({});
-                    bcv.set_options({
-                        punctuation_strategy:
-                            this.options.versionMeta?.chapterVerseSeparator === ',' ? 'eu' : 'us',
-                        invalid_passage_strategy: 'include',
-                        invalid_sequence_strategy: 'include',
-                        passage_existence_strategy: 'bc',
-                        consecutive_combination_strategy: 'separate',
-                    });
-                    context.bcv = bcv;
                 } catch (e) {
-                    Logger.warning('missing language file for bible-reference-parser', context);
+                    bcv_parser = require('bible-passage-reference-parser/cjs/en_bcv_parser').bcv_parser;
                 }
 
+                const bcv: BibleReferenceParser = new bcv_parser();
+                bcv.set_options({
+                    punctuation_strategy:
+                        this.options.versionMeta?.chapterVerseSeparator === ',' ? 'eu' : 'us',
+                    invalid_passage_strategy: 'include',
+                    invalid_sequence_strategy: 'include',
+                    passage_existence_strategy: 'bc',
+                    consecutive_combination_strategy: 'separate',
+                    non_latin_digits_strategy: 'replace',
+                    grammar: {
+                        // there seems to be a bug when matching the next-verse pattern
+                        // https://github.com/openbibleinfo/Bible-Passage-Reference-Parser/issues/33
+                        next: /^(?:\x1f\x1f\x1f)/i,
+                    }
+                });
+
+                if (this.options.bookMeta) {
+                    bcv.add_books({books: getBibleReferenceParserCustomBooks(this.options.bookMeta)});
+                }
+
+                context.bcv = bcv;
+                
                 break;
             }
             case OsisXmlNodeType.BOOK: {