+++ core/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [1.4.23]
+
+### Features
+
+-   improved cross reference detection in notes
+
 ## [1.4.20]
 
 ### Features